Patricks_updates
I did some rearranging and commented out some lines of parameters that should be set with the 1314. Setting up the Mailbox 14 for the engage 7 will be confirmed after you send me the firmware version of the engage. Power on the engage and press the center button. this will change the screen and show the firmware version on the bottom. With this version, I can provide the can object dictionary which will tell you the node ID of the engage as well as how the gauges are mapped and what information you can display.