Update Car_Testbench.txt
Some things to notice: Where you were setting the reverse speeds and accel, there was no way to come out of that setting. I moved the "else if" statement that looks for if the car is in neutral. I also added an "Else" as a default for when the car is neither in neutral nor reverse. When it is set in forward, it will be these defaults. I also noticed the speed setting you were changing, Max_Speed_SpdMx. Looking at the CPF, the controller is in Control Mode 2, Torque Mode. So I changed the Max_Speed_SpdMx to Max_Speed_TrqM. I just put a value in for the Max_Speed_TrqM and when in the Forward Else section. Adjust this to your desired setting. Also note that any time you change a max speed or accell rate in the 1314 software, that parameter value will be overwritten because you are setting it in your VCL. If you want me to show you how to create parameters that you can adjust with the 1314, let me know. It would be easier when setting up your vehicle to be able to adjust with the 1314 rather than changing the VCL, recompiling, and flashing to the controller.