Code to Light Up Buttons and Activate Noise Maker
This is code that has been verified and is working correctly on our test bench. In this version, all the button LEDs light up correctly when a button is pressed. DigOut7 is the LED for the default button and so it lights up as soon as the controller is powered, which is the same time the noise maker is activated. Buttons are also supposed to limit motor current limit. This feature will be tested with CAN mailboxes